[syntax] move the SortState out of the GeneralOptionPage
